+++ features/wallet/entity.go
@@ -5,7 +5,7 @@ import "time"
 type Core struct {
 	ID        uint
 	UserID    uint
-	Balance   float64
+	Balance   int
 	CreatedAt time.Time
 	UpdatedAt time.Time
 }
@@ -13,7 +13,7 @@ type Core struct {
 type DataInterface interface {
 	CreateWallet(id uint) error
 	GetWalletById(id uint) (Core, error)
-	UpdateBalanceMinus(id uint, amount float64) error
+	UpdateBalanceMinus(id uint, amount int) error
 }
 
 type ServiceInterface interface {
